refactor: resolve problems after rebasing
